refactoring: move common Apple code to services
